llvm: Add workaround for STLport not being a full C++11 STL
llvm checks for C++11 support in the compiler, and assumes the
STL has full C++11 support if the compiler does.
However, for the combination of gcc 4.7 and STLport, that assumption
is incorrect -- the compiler supports rvalue references just fine, but
the STL hasn't caught up yet.

Longer term, the better fix would be to fix STLport.

Credits: Bernhard Rosenkraenzer

Signed-off-by: dragon <dragon7780@hotmail.nl>